feat: support sorted comment fetching
